ComplexityInputData 소개
