+++ data/performance_result.csv
@@ -1,0 +1,80 @@
+"GPU","1048576","0","0.7170000076293945"
+"GPU","1048576","1","0.0869999974966049"
+"GPU","1048576","2","0.0780000016093254"
+"GPU","1048576","3","0.0820000022649765"
+"GPU","1048576","4","0.0820000022649765"
+"GPU","1048576","5","0.0839999988675118"
+"GPU","1048576","6","0.0829999968409538"
+"GPU","1048576","7","0.0799999982118607"
+"GPU","1048576","8","0.0769999995827675"
+"GPU","1048576","9","0.0850000008940697"
+"GPU","9437184","0","0.1230000033974648"
+"GPU","9437184","1","0.1720000058412552"
+"GPU","9437184","2","0.1650000065565109"
+"GPU","9437184","3","0.1659999936819077"
+"GPU","9437184","4","0.1650000065565109"
+"GPU","9437184","5","0.1770000010728836"
+"GPU","9437184","6","0.1659999936819077"
+"GPU","9437184","7","0.1729999929666519"
+"GPU","9437184","8","0.1739999949932098"
+"GPU","9437184","9","0.1729999929666519"
+"GPU","17825792","0","0.1749999970197678"
+"GPU","17825792","1","0.2540000081062317"
+"GPU","17825792","2","0.2520000040531158"
+"GPU","17825792","3","0.3350000083446503"
+"GPU","17825792","4","0.3070000112056732"
+"GPU","17825792","5","0.2669999897480011"
+"GPU","17825792","6","0.2590000033378601"
+"GPU","17825792","7","0.2730000019073486"
+"GPU","17825792","8","0.2540000081062317"
+"GPU","17825792","9","0.3339999914169312"
+"GPU","26214400","0","0.2240000069141388"
+"GPU","26214400","1","0.3470000028610230"
+"GPU","26214400","2","0.3420000076293945"
+"GPU","26214400","3","0.3589999973773956"
+"GPU","26214400","4","0.3479999899864197"
+"GPU","26214400","5","0.3499999940395355"
+"GPU","26214400","6","0.7089999914169312"
+"GPU","26214400","7","0.3510000109672546"
+"GPU","26214400","8","0.3490000069141388"
+"GPU","26214400","9","0.3639999926090241"
+"GPU","34603008","0","0.2860000133514404"
+"GPU","34603008","1","0.4639999866485596"
+"GPU","34603008","2","0.4300000071525574"
+"GPU","34603008","3","0.4740000069141388"
+"GPU","34603008","4","0.4839999973773956"
+"GPU","34603008","5","0.4339999854564667"
+"GPU","34603008","6","0.4210000038146973"
+"GPU","34603008","7","0.4490000009536743"
+"GPU","34603008","8","0.4469999969005585"
+"GPU","34603008","9","0.4250000119209290"
+"GPU","42991616","0","0.3140000104904175"
+"GPU","42991616","1","0.5410000085830688"
+"GPU","42991616","2","0.5139999985694885"
+"GPU","42991616","3","0.5239999890327454"
+"GPU","42991616","4","0.5360000133514404"
+"GPU","42991616","5","0.5600000023841858"
+"GPU","42991616","6","0.5329999923706055"
+"GPU","42991616","7","0.5490000247955322"
+"GPU","42991616","8","0.5360000133514404"
+"GPU","42991616","9","0.5189999938011169"
+"GPU","51380224","0","0.3569999933242798"
+"GPU","51380224","1","0.6060000061988831"
+"GPU","51380224","2","1.0440000295639038"
+"GPU","51380224","3","0.6169999837875366"
+"GPU","51380224","4","0.5920000076293945"
+"GPU","51380224","5","0.6140000224113464"
+"GPU","51380224","6","0.6190000176429749"
+"GPU","51380224","7","0.6290000081062317"
+"GPU","51380224","8","0.6299999952316284"
+"GPU","51380224","9","0.6200000047683716"
+"GPU","59768832","0","0.4029999971389771"
+"GPU","59768832","1","0.7139999866485596"
+"GPU","59768832","2","0.7039999961853027"
+"GPU","59768832","3","0.7319999933242798"
+"GPU","59768832","4","0.7039999961853027"
+"GPU","59768832","5","0.8040000200271606"
+"GPU","59768832","6","0.7110000252723694"
+"GPU","59768832","7","0.7089999914169312"
+"GPU","59768832","8","0.7110000252723694"
+"GPU","59768832","9","0.7009999752044678"
